# process-manager/src/main/webapp/openapi/openapi.yaml
openapi: 3.0.2
info:
  title: Process Platform Manager API
  description: |
    API for applications using Process Platform for running its processes; part of the API is dedicated to Process Platform worker nodes

  version: "1.0"
servers:
  - url: /process-manager/api
...
paths:
  /worker/register_node:
    post:
      tags:
        - ForWorker
      summary: Registers new node
      description: Registers a new node if not present
      requestBody:
        description: JSON with node information.
        required: true
        content:
          application/json:
            schema:
              type: string
            example: {
              "nodeId": "curatorWorker",
              "description": "Worker [curatorWorker]",
              "type": "WORKER",
              "url": "http://localhost:8089/process-worker/api/",
              "tags": [
                "testPlugin2",
                "testPlugin1-small",
                "testPlugin1-big"
              ]
            }
      responses:
        '200':
          description: Node registered or registration skipped if the node already exists
          content:
            application/json:
              schema:
                type: string
              example: {"message": "Node [curatorWorker] registered"}
        '400':
          description: Bad Request
  /worker/register_plugin:
    post:
      tags:
        - ForWorker
      summary: Registers new plugin
      description: Registers a new plugin if not present
      requestBody:
        description: JSON with plugin information.
        required: true
        content:
          application/json:
            schema:
              type: string
              example: {
                  "pluginId": "testPlugin1",
                  "description": "Testing plugin 1",
                  "mainClass": "org.ceskaexpedice.processplatform.testplugin1.TestPlugin1",
                  "payloadFieldSpecMap": {
                    "name": {
                      "type": "STRING",
                      "required": true
                    },
                    "surname": {
                      "type": "STRING",
                      "required": true
                    }
                  },
                  "scheduledProfiles": [
                    "testPlugin2"
                  ],
                  "profiles": [
                    {
                      "profileId": "testPlugin1-big",
                      "description": null,
                      "pluginId": "testPlugin1",
                      "jvmArgs": [
                        "-Xms1g",
                        "-Xmx32g"
                      ]
                    },
                    {
                      "profileId": "testPlugin1-small",
                      "description": null,
                      "pluginId": "testPlugin1",
                      "jvmArgs": [
                        "-Xms1g",
                        "-Xmx4g"
                      ]
                    }
                  ]
                }
      responses:
        '200':
          description: Plugin registered or registration skipped if the plugin already exists
          content:
            application/json:
              schema:
                type: string
              example: {"message": "Plugin [testPlugin1] registered"}
        '400':
          description: Bad Request
  /worker/next_process/{workerId}:
    get:
      tags:
        - ForWorker
      summary: Returns first process eligible for the specified worker
      description: Returns first process eligible for the specified worker. The process state is updated
      parameters:
        - name: workerId
          in: path
          required: true
          description: Registered worker id
          schema:
            type: string
          example: curatorWorker
      responses:
        '200':
          description: JSON representation of the process
          content:
            application/json:
              schema:
                type: string
                example: {
                  "processId": "ed25ce29-2149-439d-85c4-cc5e516e3036",
                  "profileId": "testPlugin1-big",
                  "pluginId": "testPlugin1",
                  "mainClass": "org.ceskaexpedice.processplatform.testplugin1.TestPlugin1",
                  "payload": {
                    "name": "Pe",
                    "surname": "Po"
                  },
                  "jvmArgs": [
                    "-Xms1g",
                    "-Xmx32g"
                  ],
                  "batchId": "ed25ce29-2149-439d-85c4-cc5e516e3036",
                  "ownerId": "PePo"
                }
        '404':
          description: No process found for the worker
  /worker/schedule_sub_process:
    post:
      tags:
        - ForWorker
      summary: Schedules new sub process
      description: Schedules new sub process in the batch
      requestBody:
        description: JSON with sub process information.
        required: true
        content:
          application/json:
            schema:
              type: string
            example: {
              "profileId" : "testPlugin2",
              "payload" : {
                "surname" : "Po",
                "name" : "Pe"
              },
              "batchId" : "6853579d-15c1-4fb9-ad11-3e107141aedb"
            }
      responses:
        '200':
          description: New sub process scheduled
          content:
            application/json:
              schema:
                type: string
              example: {"processId": "ed25ce29-2149-439d-85c4-cc5e516e3036"}
        '400':
          description: Bad Request
  /worker/pid/{processId}:
    put:
      tags:
        - ForWorker
      summary: Sets the process PID
      description: Sets the process PID
      parameters:
        - in: path
          name: processId
          description: Process id
          schema:
            type: string
          required: true

        - in: query
          name: pid
          description: PID number
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successfully updated pid
        '404':
          description: Process not found
  /worker/state/{processId}:
    put:
      tags:
        - ForWorker
      summary: Sets the process state
      description: Sets the process state
      parameters:
        - in: path
          name: processId
          description: Process id
          schema:
            type: string
          required: true

        - in: query
          name: state
          description: Process state
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successfully updated state
        '404':
          description: Process not found
  /worker/description/{processId}:
    put:
      tags:
        - ForWorker
      summary: Sets the process description
      description: Sets the process description
      parameters:
        - in: path
          name: processId
          description: Process id
          schema:
            type: string
          required: true

        - in: query
          name: description
          description: Process description
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successfully updated description
        '404':
          description: Process not found
...